+++ clients/aws/aws.go
@@ -65,6 +65,7 @@ func process(cfgObj *cfg.Config, block *hcl.Block, remainingBody hcl.Body) (cfg.
 	if diags.HasErrors() {
 		return nil, diags
 	}
+	def.DefRange = block.DefRange
 
 	clientName := block.Labels[1]