+++ clients/kafka/kafka.go
@@ -331,6 +331,7 @@ func process(config *cfg.Config, block *hcl.Block, remainingBody hcl.Body) (cfg.
 	if diags.HasErrors() {
 		return nil, diags
 	}
+	def.DefRange = block.DefRange
 
 	if len(def.Brokers) == 0 {
 		return nil, hcl.Diagnostics{{