+++ clients/openai/openai.go
@@ -48,6 +48,7 @@ func process(config *cfg.Config, block *hcl.Block, remainingBody hcl.Body) (cfg.
 	if diags.HasErrors() {
 		return nil, diags
 	}
+	clientDef.DefRange = block.DefRange
 
 	openaiCfg := openailib.DefaultConfig(clientDef.APIKey)