+++ clients/sqs/sqs_receiver.go
@@ -65,6 +65,7 @@ func processReceiver(config *cfg.Config, block *hcl.Block, remainingBody hcl.Bod
 	if diags.HasErrors() {
 		return nil, diags
 	}
+	def.DefRange = block.DefRange
 
 	clientName := block.Labels[1]