+++ clients/sqs/sqs_sender.go
@@ -66,6 +66,7 @@ func processSender(config *cfg.Config, block *hcl.Block, remainingBody hcl.Body)
 	if diags.HasErrors() {
 		return nil, diags
 	}
+	def.DefRange = block.DefRange
 
 	clientName := block.Labels[1]