+++ clients/vws/vws.go
@@ -51,6 +51,7 @@ func process(config *cfg.Config, block *hcl.Block, remainingBody hcl.Body) (cfg.
 	if diags.HasErrors() {
 		return nil, diags
 	}
+	clientDef.DefRange = block.DefRange
 
 	vinculumWebsocketsClients, ok := config.Clients["vws"]
 	if !ok {